format font when undo and redo action
