+++ apps/studio/src/lib/editor/engine/style/index.ts
@@ -8,6 +8,7 @@ import { StyleChangeType, type StyleChange } from '@onlook/models/style';
 import { makeAutoObservable, reaction } from 'mobx';
 import type { EditorEngine } from '..';
 import type { Font } from '@onlook/models/assets';
+import { convertFontString } from '@onlook/utility';
 
 export interface SelectedStyle {
     styles: Record<string, string>;
@@ -44,7 +45,33 @@ export class StyleManager {
     updateFontFamily(style: string, value: Font) {
         const styleObj = { [style]: value.id };
         const action = this.getUpdateStyleAction(styleObj);
-        this.editorEngine.action.run(action);
+        const formattedAction = {
+            ...action,
+            targets: action.targets.map((val) => ({
+                ...val,
+                change: {
+                    original: Object.fromEntries(
+                        Object.entries(val.change.original).map(([key, styleChange]) => [
+                            key,
+                            {
+                                ...styleChange,
+                                value: convertFontString(styleChange.value),
+                            },
+                        ]),
+                    ),
+                    updated: Object.fromEntries(
+                        Object.entries(val.change.updated).map(([key, styleChange]) => [
+                            key,
+                            {
+                                ...styleChange,
+                                value: convertFontString(styleChange.value),
+                            },
+                        ]),
+                    ),
+                },
+            })),
+        };
+        this.editorEngine.action.run(formattedAction);
         setTimeout(() => {
             this.editorEngine.webviews.reloadWebviews();
         }, 500);
